+++ programs/uxd/tests/integration_tests/api/program_context/mod.rs
@@ -1,11 +1,15 @@
 pub mod create_program_test_context;
 pub mod process_instruction;
-pub mod program_context;
+pub mod program_context_rpc;
+pub mod program_context_test;
+pub mod program_context_trait;
 pub mod program_error;
 pub mod read_account;
 
 pub use create_program_test_context::*;
 pub use process_instruction::*;
-pub use program_context::*;
+pub use program_context_rpc::*;
+pub use program_context_test::*;
+pub use program_context_trait::*;
 pub use program_error::*;
 pub use read_account::*;
